[ci skip] Change folder path, pre-processor name, refactor code
